Change the way of matching tests to features for some tests
