+++ Cargo.toml
@@ -74,3 +74,40 @@ default = [
     "not_like",
     "try_into",
 ]
+
+[[test]]
+name = "add_assign"
+path = "tests/add_assign.rs"
+required-features = ["add_assign_like"]
+
+[[test]]
+name = "boats_display_derive"
+path = "tests/boats_display_derive.rs"
+required-features = ["display"]
+
+[[test]]
+name = "deref_mut"
+path = "tests/deref_mut.rs"
+required-features = ["deref", "deref_mut"]
+
+
+[[test]]
+name = "no_std"
+path = "tests/no_std.rs"
+required-features = [
+    "add_assign_like",
+    "add_like",
+    "constructor",
+    "deref",
+    "deref_mut",
+    "display",
+    "from",
+    "from_str",
+    "index",
+    "index_mut",
+    "into",
+    "mul_assign_like",
+    "mul_like",
+    "not_like",
+    "try_into",
+]
