+++ tests/add_assign.rs
@@ -1,5 +1,3 @@
-#![cfg(feature = "add_assign_like")]
-
 #![allow(dead_code)]
 #[macro_use]
 extern crate derive_more;
