+++ tests/boats_display_derive.rs
@@ -1,5 +1,3 @@
-#![cfg(feature = "display")]
-
 // The following code is from https://github.com/withoutboats/display_derive/blob/232a32ee19e262aacbd2c93be5b4ce9e89a5fc30/tests/tests.rs
 // Written by without boats originally
